Fix a bug in runmany_wrap
It looks like there was a typo when the runmany wrap function was
extracted in PR #130, and it returns the function value instead of
actually executing the passed function.

It looks like it slipped through testing because the tests use the
default malt for map and ftmap, and that doesn't exercise this path. I
wrote a test derived from the use case in chef-server that revealed
the issue which seems sufficient to cover a good bit of that code.

Signed-off-by: Mark Anderson <mark@chef.io>